creating and populating a shared database

+++ clubhouse/recipes/build.rb
@@ -2,6 +2,8 @@
 # LOG NODE ATTRIBUTES
 #
 
+require 'digest'
+
 [
   "hostname", "fqdn", "domain", "ipaddress", "macaddress", "ip6address",
   "platform", "platform_version", "platform_family",
@@ -25,6 +27,38 @@ unless node.platform == "amazon"
   end
 end
 
+template "include root password in mysql configuration file for passwordless logs" do
+  path "/etc/my.cnf"
+  source "my.cnf.erb"
+  owner "root"
+  group "root"
+end
+
+members_csv_file_path = '/home/ec2-user/members.csv'
+
+cookbook_file members_csv_file_path do
+  source "members.csv"
+  action :create
+end
+
+shared_database_builder_file_path = '/home/ec2-user/create_shared_database.sql'
+
+template shared_database_builder_file_path do
+  source 'create_shared_database.sql'
+  variables ({:members_csv_file_path => members_csv_file_path})
+end
+
+if node.platform == "amazon" # todo: install mysql on non-amazon platforms
+  bash "mysql shared database setup" do
+    user 'ec2-user'
+    code <<-EOH
+      mysql -uroot -e "DROP DATABASE IF EXISTS clubhouse_db;"
+      mysql -uroot -e "CREATE DATABASE clubhouse_db;"
+      mysql -uroot clubhouse_db < #{shared_database_builder_file_path};
+    EOH
+  end
+end
+
 node["ssh_users"].each do |id, member|
   ["inbox","outbox"].each do |dir_name|
     directory "/home/#{member["name"]}/#{dir_name}" do
@@ -41,25 +75,22 @@ node["ssh_users"].each do |id, member|
     source "secret_message.erb"
     variables({
       :member_name => member["name"],
-      :passcode => "4d782a10037fbcd8cee463865cf54497", #todo: require 'digest' Digest::MD5.hexdigest("my phrase" + "1234")
-      :passphrase => "#RaiseHigh the Buff and Blue"
+      :passcode => Digest::MD5.hexdigest("my phrase" + DateTime.now.to_s),
+      :passphrase => "Raise High the Buff and Blue"
     })
   end
 
-  template "include root password in mysql configuration file for passwordless logs" do
-    path "/etc/my.cnf"
-    source "my.cnf.erb"
-    owner "root"
-    group "root"
-  end
-
-  bash "mysql setup for #{member["name"]}" do
-    user node.platform == "amazon" ? 'ec2-user' : "vagrant" # temporary workaround; todo: create users on non-amazon platforms
-    code <<-EOH
-      mysql -uroot -e "DROP DATABASE IF EXISTS #{member["name"]};"
-      mysql -uroot -e "CREATE DATABASE #{member["name"]};"
-      mysql -uroot -e "GRANT ALL ON #{member["name"]}.* TO '#{member["name"]}'@'127.0.0.1' IDENTIFIED BY '#{member["mysql_password"]}';"
-    EOH
+  if node.platform == "amazon" # todo: install mysql on non-amazon platforms
+    bash "mysql setup for #{member["name"]}" do
+      user 'ec2-user'
+      code <<-EOH
+        mysql -uroot -e "DROP DATABASE IF EXISTS #{member["name"]};"
+        mysql -uroot -e "CREATE DATABASE #{member["name"]};"
+        mysql -uroot -e "CREATE USER '#{member["name"]}'@'127.0.0.1' IDENTIFIED BY '#{member["mysql_password"]}';"
+        mysql -uroot -e "GRANT ALL ON #{member["name"]}.* TO '#{member["name"]}'@'127.0.0.1';"
+        mysql -uroot -e "GRANT SELECT ON members.* TO '#{member["name"]}'@'127.0.0.1;"
+      EOH
+    end
   end
 end
